Feat: 헤더 - 프로필 박스 로그아웃 기능 추가 및 마이페이지 삭제 #82
